+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - only lookup for properties within boundaries of memory dump.
 - only show hovered state when mouse is inside of MEMORY / PROPERTIES views.
 
+### Added
+- support to print data types Quaternion, EulerAngles, WorldPosition and 
+  WorldTransform.
+
 ------------------------
 
 ## [0.1.0] - 2024-05-06